# openapi_fr.yaml
openapi: 3.0.3
info:
  title: Wine Cellar Management API
  description: API pour gérer les utilisateurs, les permissions, les caves à vin et les bouteilles.
  version: 1.0.0
servers:
- url: http://localhost:5000
  description: Serveur de développement local

# Définition des modèles de données (schemas)
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: id
          example: "12345"
        email:
          type: string
          format: email
          example: "user@example.com"
        username:
          type: string
          example: "john_doe"
        created_at:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Unauthorized"
        message:
          type: string
          example: "Invalid token"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

# Définition des endpoints
paths:
...
  /users:
    post:
      tags:
      - Users
      summary: Crée un nouvel utilisateur
      description: Enregistre un nouvel utilisateur dans le système.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                username:
                  type: string
                  example: "john_doe"
                password:
                  type: string
                  format: password
                  example: "secure_password123"
      responses:
        201:
          description: Utilisateur créé avec succès
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        400:
          description: Requête invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      tags:
      - Users
      summary: Liste tous les utilisateurs
      description: Récupère la liste de tous les utilisateurs (réservé aux admins).
      security:
      - bearerAuth: []
      responses:
        200:
          description: Liste des utilisateurs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        403:
          description: Accès non autorisé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users/{user_id}:
    get:
      tags:
      - Users
      summary: Récupère un utilisateur
      description: Récupère les détails d'un utilisateur spécifique.
      security:
      - bearerAuth: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          format: id
        example: 12345
      responses:
        200:
          description: Détails de l'utilisateur
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        404:
          description: Utilisateur non trouvé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      tags:
      - Users
      summary: Met à jour un utilisateur
      description: Met à jour les informations d'un utilisateur.
      security:
      - bearerAuth: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          format: id
        example: 12345
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: "new_email@example.com"
                username:
                  type: string
                  example: "new_username"
                password:
                  type: string
                  format: password
                  example: "new_secure_password123"
      responses:
        200:
          description: Utilisateur mis à jour avec succès
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        404:
          description: Utilisateur non trouvé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags:
      - Users
      summary: Supprime un utilisateur
      description: Supprime un utilisateur du système.
      security:
      - bearerAuth: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          format: id
        example: 12345
      responses:
        204:
          description: Utilisateur supprimé avec succès
        404:
          description: Utilisateur non trouvé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ===== Manage Permissions =====
...